+++ tests/integ/test_data_source_api.py
@@ -78,7 +78,7 @@ from tests.resources.test_data_source_dir.test_data_source_data import (
     sql_server_create_connection_unicode_data,
     sql_server_create_connection_double_quoted_data,
 )
-from tests.utils import Utils, IS_WINDOWS
+from tests.utils import RUNNING_ON_JENKINS, Utils, IS_WINDOWS
 
 try:
     import pandas  # noqa: F401
@@ -800,6 +800,9 @@ def test_oracledb_driver_coverage(caplog):
     assert "Snowpark does not support column" in caplog.text
 
 
+@pytest.mark.skipif(
+    RUNNING_ON_JENKINS, reason="Cannot connect to oracledb from jenkins"
+)
 def test_udtf_ingestion_oracledb(session):
     from tests.parameters import ORACLEDB_CONNECTION_PARAMETERS