+++ tests/unit/scala/test_utils_suite.py
@@ -277,8 +277,10 @@ def test_zip_file_or_directory_to_stream():
             [
                 "resources/",
                 "resources/books.xml",
+                "resources/books2.xml",
                 "resources/broken.csv",
                 "resources/diamonds.csv",
+                "resources/dog.jpg",
                 "resources/fias_house.xml",
                 "resources/fias_house.large.xml",
                 "resources/iris.csv",